Refactor auth routes and split login/callback controllers
