+++ src/routes/routes.ts
@@ -2,11 +2,15 @@ import { Router } from 'express';
 import competitions from '../controllers/competitionsController';
 import rankings from "../controllers/rankingsController";
 import { teams } from '../controllers/getTeams';
+import login from '../controllers/loginController';
+import callback from '../controllers/callbackController';
 
 const router: Router = Router();
 
 router.get('/competitions', competitions);
 router.get('/rankings', rankings);
-router.get('/teams', teams)
+router.get('/teams', teams);
+router.get('/api/auth/login', login);
+router.get('/api/auth/callback', callback);
 
 export default router;
